feat : GH-19 Implement design system        Add more tests for IheNkiriTheme
